hyundai_canfd safety: use int for loop counters
Panda firmware build is freestanding C without <stddef.h>, so size_t is
unavailable. Match the canonical for-loop pattern used elsewhere in the
safety code. Cast the sizeof expression to int.

+++ opendbc/safety/modes/hyundai_canfd.h
@@ -169,7 +169,7 @@ static bool hyundai_canfd_tx_hook(const CANPacket_t *msg) {
       return false;
     }
     const uint32_t adrv_addrs[] = {0x51U, 0x160U, 0x1EAU, 0x200U, 0x345U, 0x1DAU};
-    for (size_t i = 0U; i < sizeof(adrv_addrs) / sizeof(adrv_addrs[0]); i++) {
+    for (int i = 0; i < (int)(sizeof(adrv_addrs) / sizeof(adrv_addrs[0])); i++) {
       if (msg->addr == adrv_addrs[i]) {
         return false;
       }
@@ -249,7 +249,7 @@ static bool hyundai_canfd_fwd_hook(int bus_num, int addr) {
   // LFA-steering long uses static TX-list blocking (check_relay) for 0x1A0.
   if (hyundai_longitudinal && hyundai_canfd_lka_steering) {
     const int handoff_addrs[] = {0x1a0, 0x51, 0x160, 0x1EA, 0x200, 0x345, 0x1DA};
-    for (size_t i = 0; i < sizeof(handoff_addrs) / sizeof(handoff_addrs[0]); i++) {
+    for (int i = 0; i < (int)(sizeof(handoff_addrs) / sizeof(handoff_addrs[0])); i++) {
       if (addr == handoff_addrs[i]) {
         // Without dynamic handoff: always block (preserves pre-PR static-blocking behavior).
         // With dynamic handoff: block only when openpilot is the longitudinal authority.
